changed ot bigint
git-svn-id: file:///semplestb/final_repo@973 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SEMPlestDBScripts/StoredProcs/SetBidObject.sql
@@ -14,7 +14,7 @@ CREATE PROCEDURE dbo.SetBidObject
 
 	@ProductGroupPK			INT,
 	@PromotionPK            INT,
-	@KeywordAdEngineID      INT,
+	@KeywordAdEngineID      BIGINT,
 	@Keyword				NVARCHAR(250),
 	@MicroBidAmount			INT,
 	@ApprovalStatus			VARCHAR(30) = null,
